+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `is_*` methods for `Value`, `ValueRef` and `NumberRef`.
+- Add `From<Number>`, `From<usize>` and `From<isize>` for `Value`.
 
 ## [0.1.2] - 2023-10-30